Moved algorithm registration logging message into algo method
Since the algo method that does the registering actually makes two
registration calls, moving the logging statement here prevents a
duplicate impl registration logging message for each instance
registered.

Moving the algorithm-specific logging and referencing out of the
registration structure is also good because the registration structure
does not inherently have anything to do with the algorithm system. In
fact, due to the previous logging statement, only types with the method
"type_name" could be registered with the registrar structure without a
compiler error. Referencing algorthm-specific content makes more sense
to happen in the algorithm-specific function.