Renamed ipython notebook file to kwiver, no longer config python file
Python file that used to require configuration was in regards to whether the
build-specific library directory was to be picked up. This is no longer
relevant as the code uses the LD_LIBRARY_PATH to look for directories in which
to locate the Vital C library. Since the Vital setup script normally puts the
build-tree lib directory on the LD_LIBRARY_PATH when sourced in the build
tree, the library finding function in python will find the build-tree without
the aid of the additional variable that was being configured. Thus the
variable has been removed, and no python files now require configuration.