Handle a case on the python side when an matrix dimension is 0-valued
This results in the data-array being NULL in Eigen, so we just check if
the pointer is null. If so we return a numpy array of the same zero-area
shape that the matrix is set to. There is no data memory sharing,
obviously, but that doesn't matter since there is no data...