Removed "Vital" class prefix + strengthened type checking
Added additional support for Python-side types that mask multiple
different C/C++ types due to macro expansion due to C++ generic
templating. I.e. EigenArray opaque types depend on shape and another
array cannot be initialized from the C-pointer of a different type.
Another way to explain the motivation for this is to provide
C-equivalent type checking when construcing with pointer on the Python
side.