Revised python vital object architecture
Not supporting shallow-copy options as they don't functionally make
sense. Also revised away the `from_c_pointer` class function, making it
standard practice for there to be a `from_cptr` constructor argument.
With the added abstract method `_new` con `VitalObject` implementing
classes, C-object constriction is controlled in `VitalObject`. This
separated python and c object initialization into different methods. It
is usually the case that if there is python initialization that occurs,
it needs to happen regardless of whether a new C instnace is being
constructed or not (e.g. EigenArray, Covariance, etc.).